Migrate completed exercises to numbered files
Migrate completed exercises to numbered files
