+++ learnyounode/00-hello-world.js
@@ -1,0 +1,1 @@
+console.log('HELLO WORLD');
