+++ learnyounode/01-baby-steps.js
@@ -1,0 +1,7 @@
+nums = process.argv.slice(2);
+
+function sum(a, b){
+  return +a + +b;
+}
+
+console.log(nums.reduce(sum));
