--- learnyounode/baby-steps.js
@@ -1,7 +1,0 @@
-nums = process.argv.slice(2);
-
-function sum(a, b){
-  return +a + +b;
-}
-
-console.log(nums.reduce(sum));
