--- learnyounode/hello-world.js
@@ -1,1 +1,0 @@
-console.log('HELLO WORLD');
