--- learnyounode/my-first-async-io.js
@@ -1,27 +1,0 @@
-'use strict';
-
-let fs = require('fs');
-
-let filePath = process.argv[2];
-
-let buf;
-
-function processData(data){
-  let numLines = data.split('\n').length - 1;
-  console.log(numLines);
-}
-
-function read(err, data) {
-  if (err) {
-    throw err;
-  }
-  processData(data);
-}
-
-fs.readFile(filePath, 'utf8', read);
-
-// console.log('getBuffer returns', getBuffer());
-
-// let numLines = buf.toString().split('\n').length - 1;
-//
-// console.log(numLines);
